# [improve][pip] PIP-294: Filter out the bundles with low throughput in Bundle load report