@Demogorgon314 2023-08-14
This is an excellent optimization! But we might need to add a PIP for this feature, since it introduces a new configuration.

We can also reuse the configuration like the `ExtensibleLoadManager`? It only uploads topK hot bundle data to ZK.

https://github.com/apache/pulsar/blob/master/pulsar-broker/src/main/java/org/apache/pulsar/broker/loadbalance/extensions/reporter/TopBundleLoadDataReporter.java#L75