@thetumbled 2023-08-14
> This is an excellent optimization! But we might need to add a PIP for this feature, since it introduces a new configuration.
> 
> We can also reuse the configuration like the `ExtensibleLoadManager`? It only uploads topK hot bundle data to ZK.
> 
> https://github.com/apache/pulsar/blob/master/pulsar-broker/src/main/java/org/apache/pulsar/broker/loadbalance/extensions/reporter/TopBundleLoadDataReporter.java#L75

Ok, i will create a PIP for this feature later. 
I think that limiting the bundle count to be exported by msg rate or throughput is better than the number of bundles. As the number of bundles and the throughput in the cluster changes dynamically, users can't select a reasonable number of bundles easily, while based on throughput, we can be sure that bundles with throughput less than 0.1M is useless for loadbalance no matter how the cluster change.