@thetumbled 2023-08-17
I have open a PIP for this feature, please help to review, thanks. @codelipenghui @heesung-sn @Demogorgon314 @congbobo184 
https://github.com/apache/pulsar/pull/21011
https://lists.apache.org/thread/nsty5pbwl5lqflfjb7sd44g3h5tpcrpn