@thetumbled 2023-08-18
> > the bundle count to be exported by msg rate or throughput is better than the number of bundles. As the number of bundles and the throughput in the cluster changes dynamically, users can't select a reasonable number of bundles easily, while based on throughput, we can be sure that bundles with throughput less than 0.1M is useless for loadbalance no matter how the cluster change.
> 
> It would be great if we could also add this throughput-based bundle report size control in the ExtensibleLoadManager.

I have implemented this feature in the `ExtensibleLoadManager`.
PTAL, thanks. @Demogorgon314 @heesung-sn